requirements.txt
streamlit
streamlit-code-editor
duckdb